script/misc: toggle "ssfx_is_underground" and rain drops control dynamically
OGSR engine expects this to pass it to the shaders.
"ssfx_is_underground" does something with the lightning and should be
turned on when on an indoor level.
Rain drop control should be turned on only when you have some glass on
your face. This can be a gasmask/helmet or a night vision. Pass their
status from AMK and SA code and switch the config when anything changes.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>