all.spawn: add missing stash box near the start
During the playtesting, I found that it's currently impossible to
proceed after the game starts due to that you don't have any items
to cross the bridge :D
Follow the original NS2009 path and spawn a stash with a vodka and
a medkit on the abandoned farm near the start. But still don't give
any weapon to the player as the corpse near the stash always has
one -- you only need to avoid couple fleshes.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>